fix project errors
update material.cpp paths

+++ Season_Shift/Season_Shift/Graphics/GfxResourceDevice.cpp
@@ -19,8 +19,8 @@ std::shared_ptr<DXTexture> GfxResourceDevice::loadTexture(std::string filepath)
 		DXTexture::Type::TEX2D,
 	};
 	desc.desc2D = {
-		texData.width,
-		texData.height,
+		static_cast<UINT>(texData.width),
+		static_cast<UINT>(texData.height),
 		0,
 		1,
 		DXGI_FORMAT_R8G8B8A8_UNORM_SRGB,
